FastMPOContractions

Documentation for FastMPOContractions.

FastMPOContractions.ReducedFitProblem

FastMPOContractions.contract_fit

FastMPOContractions.error_contract

FastMPOContractions.ReducedFitProblem — Type

A ReducedFitProblem represents the projection of an MPS input_state onto the basis of a different MPS state. state may be an approximation of input_state.

     *--*--*-      -*--*--*--*--*--* <state|
     |  |  |  |  |  |  |  |  |  |  |
     o--o--o-      -o--o--o--o--o--o |input_state>

FastMPOContractions.contract_fit — Method

Contract M1 and M2, and return the result as an MPO.

FastMPOContractions.error_contract — Method

Taken from from ITensors.jl with some modifications.

The dominator is replaced by <y|A|x> instead of <x|A|A|x> to reduce the computational cost.

Specialized to MPO-MPO contraction.

error_contract(y::MPS, A::MPO, x::MPS;
               make_inds_match::Bool = true)
error_contract(y::MPS, x::MPS, x::MPO;
               make_inds_match::Bool = true)

Compute the distance between A|x> and an approximation MPS y: | |y> - A|x> |/| A|x> | = √(1 + (<y|y> - 2*real(<y|A|x>))/<y|A|x>).

If make_inds_match = true, the function attempts match the site indices of y with the site indices of A that are not common with x.